backend/pyproject.toml
[project]
name = "pokemon-analysis"
version = "0.1.0"
description = "Serviço de análise de dados de Pokémon"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastmcp>=0.1.0",
    "pandas>=1.3.0",
    "numpy==1.24.3",
    "matplotlib==3.7.1",
    "seaborn==0.12.2",
    "requests>=2.26.0",
    "aiohttp>=3.8.0",
    "python-dotenv>=0.19.0",
    "uvicorn>=0.15.0",
    "starlette>=0.14.0",
    "sse-starlette>=1.0.0",
]
